Refactor settings component
